Refactor mi, mn, ms into common and chtml components
